[2063] Refactor Observer library code back to cmd/       Author: Miklos Szegedi
