+++ cmd/observer/node_builder/observer_builder.go
@@ -1,4 +1,4 @@
-package observer
+package node_builder
 
 import (
 	"encoding/json"
