+++ cmd/observer/node_builder/upstream_connector.go
@@ -1,4 +1,4 @@
-package observer
+package node_builder
 
 import (
 	"context"
